requirements.txt
django 
djangorestframework==3.14.0
django-rest-knox==4.2.0
drf-yasg